Cleanup each state, organize parameters, things are working in sim
